experiments/gepa_gemma3/requirements.txt
dspy-ai>=2.5.1
rapidfuzz>=3.9.0
python-dotenv>=1.0.1
transformers>=4.46.0
torch>=2.5.1
sentencepiece>=0.2.0
google-genai>=0.2.0
jsonschema>=4.23.0
